+++ backend/canisters/group/impl/src/updates/send_message.rs
@@ -119,6 +119,7 @@ pub(crate) fn send_message_impl(
         },
         finalised,
         args.og_previews,
+        &state.data.migrated_user_ids,
         now,
     )?;
 
@@ -183,6 +184,7 @@ pub(crate) fn send_message_with_completed_transfer(
         },
         true,
         args.og_previews,
+        &state.data.migrated_user_ids,
         now,
     )?;